test/test_circuit_breaker.py
import unittest
from src.safety.circuit_breaker import CircuitBreaker

class TestCircuitBreaker(unittest.TestCase):

    def setUp(self):
        self.circuit_breaker = CircuitBreaker(risk_threshold=0.8)

    def test_initialization(self):
        self.assertEqual(self.circuit_breaker.risk_threshold, 0.8)

    def test_check_risk_below_threshold(self):
        result = self.circuit_breaker.check_risk(0.5)
        self.assertFalse(result)

    def test_check_risk_equal_threshold(self):
        result = self.circuit_breaker.check_risk(0.8)
        self.assertFalse(result)

    def test_check_risk_above_threshold(self):
        result = self.circuit_breaker.check_risk(0.9)
        self.assertTrue(result)

    def test_trigger_circuit_breaker(self):
        # Assuming trigger_circuit_breaker sets an internal state or flag
        self.circuit_breaker.trigger_circuit_breaker()
        # Add assertions if there's any state change to verify

    def test_reset_circuit_breaker(self):
        # Assuming reset_circuit_breaker resets internal state
        self.circuit_breaker.reset_circuit_breaker()
        # Add assertions if there's any state change to verify

if __name__ == '__main__':
    unittest.main()
